% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_acquisition.R
\name{acquire_oai_data}
\alias{acquire_oai_data}
\title{Acquire Osteoarthritis Initiative (OAI) Data}
\usage{
acquire_oai_data(data_type = "clinical", visit_number = 0)
}
\arguments{
\item{data_type}{Type of OAI data ("clinical", "imaging", "biomarker")}

\item{visit_number}{Visit number (0-96 months)}
}
\value{
Processed OAI dataset
}
\description{
Downloads and processes OAI clinical data for longitudinal
osteoarthritis progression modeling tutorials.
}
