% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_analytics.R
\name{analyze_cluster_characteristics}
\alias{analyze_cluster_characteristics}
\title{Analyze Cluster Characteristics}
\usage{
analyze_cluster_characteristics(original_data, clusters, numeric_data)
}
\arguments{
\item{original_data}{Original patient data}

\item{clusters}{Cluster assignments}

\item{numeric_data}{Numeric data used for clustering}
}
\value{
Cluster characteristics
}
\description{
Analyze Cluster Characteristics
}
