% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_integration.R, R/uncertainty_analysis.R
\name{analyze_parameter_correlations}
\alias{analyze_parameter_correlations}
\title{Analyze Parameter Correlations}
\usage{
analyze_parameter_correlations(psa_results, outcome_var = "icer")

analyze_parameter_correlations(psa_results, outcome_var = "icer")
}
\arguments{
\item{psa_results}{PSA results}

\item{outcome_var}{Outcome variable to analyze}
}
\value{
Parameter correlation analysis

Parameter correlation analysis
}
\description{
Analyze Parameter Correlations

Analyze Parameter Correlations
}
