% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_integration.R, R/uncertainty_analysis.R
\name{analyze_parameter_influence}
\alias{analyze_parameter_influence}
\title{Analyze Parameter Influence (Tornado Diagram Data)}
\usage{
analyze_parameter_influence(psa_results, outcome_var = "icer", n_levels = 5)

analyze_parameter_influence(psa_results, outcome_var = "icer", n_levels = 5)
}
\arguments{
\item{psa_results}{PSA results}

\item{outcome_var}{Outcome variable to analyze ("total_cost", "total_qaly", or "icer")}

\item{n_levels}{Number of parameter levels for analysis}
}
\value{
Tornado diagram data

Tornado diagram data
}
\description{
Analyze Parameter Influence (Tornado Diagram Data)

Analyze Parameter Influence (Tornado Diagram Data)
}
