% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_interventions_fcn.R
\name{apply_bmi_modification}
\alias{apply_bmi_modification}
\title{Apply a BMI Modification Intervention}
\usage{
apply_bmi_modification(am, intervention)
}
\arguments{
\item{am}{The attribute matrix.}

\item{intervention}{The intervention to apply.}
}
\value{
The updated attribute matrix.
}
\description{
Apply a BMI Modification Intervention
}
