% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_interventions_fcn.R
\name{apply_interventions}
\alias{apply_interventions}
\title{Apply Interventions to the Population}
\usage{
apply_interventions(attribute_matrix, intervention_params, year)
}
\arguments{
\item{attribute_matrix}{The main data frame of the simulation population.}

\item{intervention_params}{A list of parameters for the interventions.}

\item{year}{The current simulation year.}
}
\value{
The updated attribute_matrix after applying the interventions.
}
\description{
This function applies one or more interventions to the attribute matrix.
Interventions are defined in a configuration file and can modify model
parameters or directly affect individuals' attributes.
}
