% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_policy_levers_fcn.R
\name{apply_policy_levers}
\alias{apply_policy_levers}
\title{Apply Policy Levers to Model Parameters}
\usage{
apply_policy_levers(params, policy_levers)
}
\arguments{
\item{params}{A list containing all model parameters (e.g., the output of \code{load_config}).}

\item{policy_levers}{A list of policy lever objects from the simulation configuration.}
}
\value{
The modified list of parameters.
}
\description{
This function modifies a list of model parameters (including coefficients,
costs, and utilities) based on a set of enabled policy levers defined in the
simulation configuration.
}
\examples{
# params <- load_config("config/coefficients.yaml")
# sim_config <- load_config("config/simulation.yaml")
# modified_params <- apply_policy_levers(params, sim_config$policy_levers)
}
