% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_interventions_fcn.R
\name{apply_qaly_cost_modification}
\alias{apply_qaly_cost_modification}
\title{Apply a QALY and Cost Modification Intervention}
\usage{
apply_qaly_cost_modification(am, intervention)
}
\arguments{
\item{am}{The attribute matrix.}

\item{intervention}{The intervention to apply.}
}
\value{
The updated attribute matrix.
}
\description{
Apply a QALY and Cost Modification Intervention
}
