% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_interventions_fcn.R
\name{apply_tka_risk_modification}
\alias{apply_tka_risk_modification}
\title{Apply a TKA Risk Modification Intervention}
\usage{
apply_tka_risk_modification(am, intervention)
}
\arguments{
\item{am}{The attribute matrix.}

\item{intervention}{The intervention to apply.}
}
\value{
The updated attribute matrix.
}
\description{
Apply a TKA Risk Modification Intervention
}
