% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_modeling.R
\name{assess_calibration}
\alias{assess_calibration}
\title{Assess Model Calibration}
\usage{
assess_calibration(predictions, actual_outcomes)
}
\arguments{
\item{predictions}{Model predictions}

\item{actual_outcomes}{Actual outcomes}
}
\value{
Calibration assessment
}
\description{
Assess Model Calibration
}
