% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_integration.R
\name{assess_psa_convergence}
\alias{assess_psa_convergence}
\title{Assess PSA Convergence}
\usage{
assess_psa_convergence(psa_results, psa_params)
}
\arguments{
\item{psa_results}{PSA results from framework}

\item{psa_params}{PSA control parameters}
}
\value{
Convergence assessment results
}
\description{
Assess PSA Convergence
}
