% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_calibration.R
\name{bayesian_model_calibration}
\alias{bayesian_model_calibration}
\title{Bayesian Model Calibration}
\usage{
bayesian_model_calibration(bayesian_model, validation_data, config)
}
\arguments{
\item{bayesian_model}{Fitted Bayesian model}

\item{validation_data}{Validation dataset}

\item{config}{Configuration}
}
\value{
Calibration results
}
\description{
Bayesian Model Calibration
}
