% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_calibration.R
\name{bayesian_parameter_estimation}
\alias{bayesian_parameter_estimation}
\title{Bayesian Parameter Estimation for Complication Risks}
\usage{
bayesian_parameter_estimation(data, outcome_var, config)
}
\arguments{
\item{data}{Patient data with outcomes}

\item{outcome_var}{Outcome variable to model}

\item{config}{Bayesian configuration}
}
\value{
Bayesian model results
}
\description{
Bayesian Parameter Estimation for Complication Risks
}
