% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_monitoring.R
\name{benchmark_core_functions}
\alias{benchmark_core_functions}
\title{Benchmark core functions of the AUS-OA package}
\usage{
benchmark_core_functions(iterations = 5)
}
\arguments{
\item{iterations}{Number of benchmarking iterations to run}
}
\value{
Benchmark results for all tested functions
}
\description{
Benchmark core functions of the AUS-OA package
}
