% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bmi_mod_fcn}
\alias{bmi_mod_fcn}
\title{Update BMI for one cycle (Optimized with data.table)}
\usage{
bmi_mod_fcn(am_curr, cycle_coefficients, bmi_customisations)
}
\arguments{
\item{am_curr}{A data.table representing the attribute matrix for the current
cycle.}

\item{cycle_coefficients}{A list containing the regression coefficients.}

\item{bmi_customisations}{A data.frame containing calibration factors.}
}
\value{
The updated attribute matrix (a copy) with a new 'd_bmi' column.
}
\description{
This function updates the Body Mass Index (BMI) for each individual in the
attribute matrix for a single simulation cycle. This version is optimized
for performance using the data.table package.
}
