% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats_temp_fcn.R
\name{BMI_summary_data}
\alias{BMI_summary_data}
\title{Summarise BMI Data}
\usage{
BMI_summary_data(am_all)
}
\arguments{
\item{am_all}{A data.frame representing the full simulation output, containing
data for all individuals over all cycles.}
}
\value{
A data.frame with the summarised BMI statistics.
}
\description{
Processes the full simulation output to calculate the proportion of the
population that is overweight or obese, stratified by year, age category,
and sex.
}
