% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_framework.R
\name{calculate_convergence_diagnostics}
\alias{calculate_convergence_diagnostics}
\title{Calculate Convergence Diagnostics}
\usage{
calculate_convergence_diagnostics(psa_results)
}
\arguments{
\item{psa_results}{Results from PSA simulation}
}
\value{
List of convergence diagnostics
}
\description{
Calculate Convergence Diagnostics
}
