% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_optimization.R
\name{calculate_costs_efficient}
\alias{calculate_costs_efficient}
\title{Memory-Efficient Cost Calculator}
\usage{
calculate_costs_efficient(dt, cost_params)
}
\arguments{
\item{dt}{A data table containing the population data}

\item{cost_params}{A list of cost parameters}
}
\value{
Updated data table with cost columns added
}
\description{
Calculates costs using memory-efficient operations.
}
