% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_private_pathways_module.R
\name{calculate_equity_metrics}
\alias{calculate_equity_metrics}
\title{Calculate Equity and Access Metrics}
\usage{
calculate_equity_metrics(am_curr)
}
\arguments{
\item{am_curr}{Current attribute matrix}
}
\value{
Summary statistics on equity and access
}
\description{
Calculate Equity and Access Metrics
}
