% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceac_enhanced.R
\name{calculate_nmb}
\alias{calculate_nmb}
\title{Calculate Net Monetary Benefit (NMB)}
\usage{
calculate_nmb(psa_results, wtp_threshold)
}
\arguments{
\item{psa_results}{PSA results}

\item{wtp_threshold}{Willingness-to-pay threshold}
}
\value{
NMB statistics
}
\description{
Calculate Net Monetary Benefit (NMB)
}
