% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_private_pathways_module.R
\name{calculate_pathway_costs}
\alias{calculate_pathway_costs}
\title{Calculate Pathway-Specific Costs}
\usage{
calculate_pathway_costs(am_curr, cost_coefficients)
}
\arguments{
\item{am_curr}{Current attribute matrix}

\item{cost_coefficients}{Parameters for pathway-specific costs}
}
\value{
Updated attribute matrix with detailed cost breakdowns
}
\description{
Calculate Pathway-Specific Costs
}
