% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_analytics.R
\name{calculate_pattern_characteristics}
\alias{calculate_pattern_characteristics}
\title{Calculate Pattern Characteristics}
\usage{
calculate_pattern_characteristics(outcome, time)
}
\arguments{
\item{outcome}{Outcome values}

\item{time}{Time points}
}
\value{
Pattern characteristics
}
\description{
Calculate Pattern Characteristics
}
