% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pji_module.R
\name{calculate_pji_impacts}
\alias{calculate_pji_impacts}
\title{Calculate PJI Costs and QALY Impacts}
\usage{
calculate_pji_impacts(am_curr, pji_cost_params, pji_qaly_params)
}
\arguments{
\item{am_curr}{Current attribute matrix}

\item{pji_cost_params}{Cost parameters}

\item{pji_qaly_params}{QALY parameters}
}
\value{
List with cost and QALY impacts
}
\description{
Calculate PJI Costs and QALY Impacts
}
