% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_modeling.R
\name{calculate_prediction_uncertainty}
\alias{calculate_prediction_uncertainty}
\title{Calculate Prediction Uncertainty}
\usage{
calculate_prediction_uncertainty(predictions, actual_outcomes = NULL)
}
\arguments{
\item{predictions}{List of model predictions}

\item{actual_outcomes}{Actual outcomes (if available)}
}
\value{
Uncertainty statistics
}
\description{
Calculate Prediction Uncertainty
}
