% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_qaly_fcn.R
\name{calculate_qaly}
\alias{calculate_qaly}
\title{Calculate Quality-Adjusted Life Years (QALYs)}
\usage{
calculate_qaly(attribute_matrix, utility_params)
}
\arguments{
\item{attribute_matrix}{The main data frame of the simulation population.}

\item{utility_params}{A list of parameters defining the utility decrements
for various health states.}
}
\value{
The updated attribute_matrix with the \code{d_sf6d} column populated with
the total utility decrement for the cycle.
}
\description{
This function calculates the health utility (SF-6D) for each individual
based on their current health state, including KL grade, BMI, and
comorbidities.
}
