% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_analytics.R
\name{calculate_trend}
\alias{calculate_trend}
\title{Calculate Trend in Trajectory}
\usage{
calculate_trend(outcome, time)
}
\arguments{
\item{outcome}{Outcome values}

\item{time}{Time points}
}
\value{
Trend analysis
}
\description{
Calculate Trend in Trajectory
}
