% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceac_enhanced.R
\name{calculate_voi}
\alias{calculate_voi}
\title{Calculate Value of Information (VOI)}
\usage{
calculate_voi(psa_results, wtp_threshold)
}
\arguments{
\item{psa_results}{PSA results}

\item{wtp_threshold}{Willingness-to-pay threshold}
}
\value{
VOI statistics
}
\description{
Calculate Value of Information (VOI)
}
