% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waiting_list_module.R
\name{calculate_wait_time_impacts}
\alias{calculate_wait_time_impacts}
\title{Calculate Wait Time Impacts on Outcomes}
\usage{
calculate_wait_time_impacts(am_curr, wait_time_coefficients)
}
\arguments{
\item{am_curr}{Current attribute matrix with wait times}

\item{wait_time_coefficients}{Parameters for wait time impacts}
}
\value{
Updated attribute matrix with wait time impacts
}
\description{
Calculate Wait Time Impacts on Outcomes
}
