% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_monitoring.R
\name{compare_function_performance}
\alias{compare_function_performance}
\title{Compare performance between function versions}
\usage{
compare_function_performance(func1, func2, data, iterations = 5)
}
\arguments{
\item{func1}{First function to compare}

\item{func2}{Second function to compare}

\item{data}{Input data for both functions}

\item{iterations}{Number of iterations to run}
}
\value{
Comparison results
}
\description{
Compare performance between function versions
}
