% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_framework.R
\name{create_config_from_parameters}
\alias{create_config_from_parameters}
\title{Create Configuration from Parameter Vector}
\usage{
create_config_from_parameters(params)
}
\arguments{
\item{params}{Vector of parameter values}
}
\value{
Configuration list suitable for simulation
}
\description{
Create Configuration from Parameter Vector
}
