% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcement_learning.R
\name{create_initial_state_distribution}
\alias{create_initial_state_distribution}
\title{Create Initial State Distribution}
\usage{
create_initial_state_distribution(patient_data)
}
\arguments{
\item{patient_data}{Patient data}
}
\value{
Initial state probabilities
}
\description{
Create Initial State Distribution
}
