% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_framework.R
\name{create_model_ensemble}
\alias{create_model_ensemble}
\title{Create Model Ensemble}
\usage{
create_model_ensemble(models, features, outcome)
}
\arguments{
\item{models}{List of trained models}

\item{features}{Feature matrix}

\item{outcome}{Outcome vector}
}
\value{
Ensemble model
}
\description{
Create Model Ensemble
}
