% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcement_learning.R
\name{create_reward_function}
\alias{create_reward_function}
\title{Create Reward Function}
\usage{
create_reward_function(patient_data, config)
}
\arguments{
\item{patient_data}{Patient data}

\item{config}{Configuration}
}
\value{
Reward matrix
}
\description{
Create Reward Function
}
