% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_modeling.R
\name{create_risk_stratification}
\alias{create_risk_stratification}
\title{Create Risk Stratification Groups}
\usage{
create_risk_stratification(predictions, n_groups = 3)
}
\arguments{
\item{predictions}{Risk predictions}

\item{n_groups}{Number of risk groups}
}
\value{
Risk stratification results
}
\description{
Create Risk Stratification Groups
}
