% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcement_learning.R
\name{create_transition_function}
\alias{create_transition_function}
\title{Create State Transition Function}
\usage{
create_transition_function(patient_data, config)
}
\arguments{
\item{patient_data}{Patient data}

\item{config}{Configuration}
}
\value{
Transition probability matrix
}
\description{
Create State Transition Function
}
