% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcement_learning.R
\name{create_weighted_reward_function}
\alias{create_weighted_reward_function}
\title{Create Weighted Reward Function}
\usage{
create_weighted_reward_function(environment, config, weights)
}
\arguments{
\item{environment}{Environment}

\item{config}{Configuration}

\item{weights}{Objective weights}
}
\value{
Weighted reward function
}
\description{
Create Weighted Reward Function
}
