% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcement_learning.R
\name{define_treatment_environment}
\alias{define_treatment_environment}
\title{Define Treatment Optimization Environment}
\usage{
define_treatment_environment(patient_data, config)
}
\arguments{
\item{patient_data}{Patient data}

\item{config}{RL configuration}
}
\value{
Treatment environment
}
\description{
Define Treatment Optimization Environment
}
