% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvt_module.R
\name{dvt_module}
\alias{dvt_module}
\title{Main DVT Module Function}
\usage{
dvt_module(am_curr, am_new, dvt_params)
}
\arguments{
\item{am_curr}{Current attribute matrix}

\item{am_new}{Next cycle attribute matrix}

\item{dvt_params}{DVT parameters from configuration}
}
\value{
List containing updated matrices and DVT summary
}
\description{
Main DVT Module Function
}
