% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats_temp_fcn.R
\name{f_get_percent_N_from_binary}
\alias{f_get_percent_N_from_binary}
\title{Calculates the percentage and frequency of binary (0/1) variables for each
level of specified grouping variables.}
\usage{
f_get_percent_N_from_binary(df, group_vars)
}
\arguments{
\item{df}{A data.frame containing the data.}

\item{group_vars}{A character vector of variable names to group by.}
}
\value{
A data.frame with summary statistics (percentage and frequency) for
each binary variable and group combination.
}
\description{
Calculates the percentage and frequency of binary (0/1) variables for each
level of specified grouping variables.
}
