% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions_fcn.R
\name{f_plot_distribution}
\alias{f_plot_distribution}
\title{Plot Distribution by Age and Sex for a Given Year}
\usage{
f_plot_distribution(data, variable, yearv)
}
\arguments{
\item{data}{A data.frame containing the data to plot.}

\item{variable}{The variable (unquoted) to be plotted on the y-axis.}

\item{yearv}{The specific year to filter the data by.}
}
\value{
A ggplot object.
}
\description{
Generates a bar chart showing the distribution of a variable across age groups
and by sex for a specific year.
}
