% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions_fcn.R
\name{f_plot_trend_overall}
\alias{f_plot_trend_overall}
\title{Plot Overall Time Series Trend}
\usage{
f_plot_trend_overall(
  data,
  x_var,
  y_var,
  shape_var,
  color_var,
  colors,
  x_label,
  y_label
)
}
\arguments{
\item{data}{A data.frame containing the data to plot.}

\item{x_var}{A character string naming the x-axis variable (e.g., "year").}

\item{y_var}{A character string naming the y-axis variable.}

\item{shape_var}{A character string naming the variable for point shape.}

\item{color_var}{A character string naming the variable for point/line color.}

\item{colors}{A vector of color values to use for the plot.}

\item{x_label}{A character string for the x-axis label.}

\item{y_label}{A character string for the y-axis label.}
}
\value{
A ggplot object.
}
\description{
Generates a time series plot for a given variable, showing the overall
trend for the entire population.
}
