% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_analytics.R
\name{find_optimal_clusters_elbow}
\alias{find_optimal_clusters_elbow}
\title{Find Optimal Number of Clusters (Elbow Method)}
\usage{
find_optimal_clusters_elbow(wss, k_range)
}
\arguments{
\item{wss}{Within-cluster sum of squares}

\item{k_range}{Range of k values}
}
\value{
Optimal number of clusters
}
\description{
Find Optimal Number of Clusters (Elbow Method)
}
