% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_analytics.R
\name{find_optimal_clusters_silhouette}
\alias{find_optimal_clusters_silhouette}
\title{Find Optimal Number of Clusters (Silhouette Method)}
\usage{
find_optimal_clusters_silhouette(data, k_range)
}
\arguments{
\item{data}{Scaled data matrix}

\item{k_range}{Range of k values}
}
\value{
Optimal number of clusters
}
\description{
Find Optimal Number of Clusters (Silhouette Method)
}
