% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_integration.R
\name{generate_ceac}
\alias{generate_ceac}
\title{Generate Cost-Effectiveness Acceptability Curve (CEAC)}
\usage{
generate_ceac(psa_results, wtp_threshold = NULL, wtp_range = NULL)
}
\arguments{
\item{psa_results}{PSA results}

\item{wtp_threshold}{Willingness-to-pay threshold}

\item{wtp_range}{Range of WTP values for CEAC}
}
\value{
CEAC data for plotting
}
\description{
Generate Cost-Effectiveness Acceptability Curve (CEAC)
}
