% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceac_enhanced.R
\name{generate_ceac_bootstrap}
\alias{generate_ceac_bootstrap}
\title{Generate CEAC with Bootstrap Confidence Intervals}
\usage{
generate_ceac_bootstrap(
  psa_results,
  wtp_threshold = NULL,
  wtp_range = NULL,
  n_bootstrap = 1000,
  conf_level = 0.95
)
}
\arguments{
\item{psa_results}{PSA results}

\item{wtp_threshold}{Willingness-to-pay threshold}

\item{wtp_range}{Range of WTP values for CEAC}

\item{n_bootstrap}{Number of bootstrap samples}

\item{conf_level}{Confidence level for intervals}
}
\value{
CEAC data with bootstrap confidence intervals
}
\description{
Generate CEAC with Bootstrap Confidence Intervals
}
