% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceac_enhanced.R
\name{generate_enhanced_ceac}
\alias{generate_enhanced_ceac}
\title{Generate Enhanced CEAC Analysis}
\usage{
generate_enhanced_ceac(
  psa_results,
  wtp_threshold = 50000,
  include_bootstrap = TRUE,
  n_bootstrap = 1000
)
}
\arguments{
\item{psa_results}{PSA results}

\item{wtp_threshold}{Willingness-to-pay threshold}

\item{include_bootstrap}{Include bootstrap confidence intervals}

\item{n_bootstrap}{Number of bootstrap samples}
}
\value{
Enhanced CEAC analysis
}
\description{
Generate Enhanced CEAC Analysis
}
