% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_framework.R
\name{generate_psa_summary}
\alias{generate_psa_summary}
\title{Generate PSA Summary Report}
\usage{
generate_psa_summary(psa_results)
}
\arguments{
\item{psa_results}{Complete PSA results}
}
\value{
Formatted summary report
}
\description{
Generate PSA Summary Report
}
