% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_integration.R
\name{get_default_psa_parameters}
\alias{get_default_psa_parameters}
\title{Get Default PSA Parameters}
\usage{
get_default_psa_parameters()
}
\value{
Default PSA parameter configuration
}
\description{
Get Default PSA Parameters
}
