% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_processing.R
\name{get_optimal_cores}
\alias{get_optimal_cores}
\title{Get Recommended Number of Cores}
\usage{
get_optimal_cores(task_type = "cpu_intensive", max_cores = NULL)
}
\arguments{
\item{task_type}{Type of task ("cpu_intensive", "io_intensive", "memory_limited")}

\item{max_cores}{Maximum cores to recommend (NULL for automatic)}
}
\value{
Recommended number of cores to use
}
\description{
Returns a recommended number of cores based on system resources and task type.
}
