% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_param_value}
\alias{get_param_value}
\title{Get a parameter value from the simulation setup}
\usage{
get_param_value(param_name)
}
\arguments{
\item{param_name}{The name of the parameter to get.}
}
\value{
The value of the parameter.
}
\description{
Get a parameter value from the simulation setup
}
