% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_interventions_fcn.R
\name{get_target_indices}
\alias{get_target_indices}
\title{Get Indices of the Target Population for an Intervention}
\usage{
get_target_indices(am, target_def)
}
\arguments{
\item{am}{The attribute matrix.}

\item{target_def}{A list defining the target population criteria.}
}
\value{
A logical vector indicating which rows of the attribute matrix
  belong to the target population.
}
\description{
Get Indices of the Target Population for an Intervention
}
