% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_kl_grades_fcn.R
\name{initialize_kl_grades}
\alias{initialize_kl_grades}
\title{Initialize KL Grades}
\usage{
initialize_kl_grades(am, utilities, initial_kl_grades)
}
\arguments{
\item{am}{A data.frame of the attribute matrix.}

\item{utilities}{A list of utility coefficients.}

\item{initial_kl_grades}{A list of initial KL grade probabilities.}
}
\value{
A data.frame with the attribute matrix updated with KL grades.
}
\description{
This function initializes the Kellgren-Lawrence (KL) grades for the population
based on their osteoarthritis (OA) status.
}
