% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_optimization.R
\name{manage_memory}
\alias{manage_memory}
\title{Garbage Collection Helper}
\usage{
manage_memory(verbose = TRUE)
}
\arguments{
\item{verbose}{Whether to print memory information}
}
\value{
List with memory information before and after GC
}
\description{
Performs targeted garbage collection and reports memory usage.
}
