% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvt_module.R
\name{model_dvt_treatment}
\alias{model_dvt_treatment}
\title{Model DVT Treatment and Outcomes}
\usage{
model_dvt_treatment(am_curr, dvt_coefficients)
}
\arguments{
\item{am_curr}{Current attribute matrix with DVT events}

\item{dvt_coefficients}{DVT treatment coefficients}
}
\value{
Updated attribute matrix with treatment outcomes
}
\description{
Model DVT Treatment and Outcomes
}
