% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waiting_list_module.R
\name{model_pathway_selection}
\alias{model_pathway_selection}
\title{Model Public vs Private Pathway Selection}
\usage{
model_pathway_selection(am_curr, pathway_coefficients)
}
\arguments{
\item{am_curr}{Current attribute matrix}

\item{pathway_coefficients}{Parameters for pathway selection}
}
\value{
Updated attribute matrix with pathway assignments
}
\description{
Model Public vs Private Pathway Selection
}
