% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_private_pathways_module.R
\name{model_patient_satisfaction}
\alias{model_patient_satisfaction}
\title{Model Patient Satisfaction and Experience}
\usage{
model_patient_satisfaction(am_curr, satisfaction_coefficients)
}
\arguments{
\item{am_curr}{Current attribute matrix}

\item{satisfaction_coefficients}{Parameters for patient experience}
}
\value{
Updated attribute matrix with satisfaction metrics
}
\description{
Model Patient Satisfaction and Experience
}
