% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pji_module.R
\name{model_pji_treatment}
\alias{model_pji_treatment}
\title{Model PJI Treatment Pathways}
\usage{
model_pji_treatment(am_curr, pji_treatment_params)
}
\arguments{
\item{am_curr}{Current attribute matrix}

\item{pji_treatment_params}{Treatment parameters}
}
\value{
Updated attribute matrix with treatment outcomes
}
\description{
Model PJI Treatment Pathways
}
