% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcement_learning.R
\name{multi_objective_optimization}
\alias{multi_objective_optimization}
\title{Multi-Objective Treatment Optimization}
\usage{
multi_objective_optimization(environment, config)
}
\arguments{
\item{environment}{Treatment environment}

\item{config}{RL configuration}
}
\value{
Pareto optimal policies
}
\description{
Multi-Objective Treatment Optimization
}
