% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats_temp_fcn.R
\name{OA_summary_fcn}
\alias{OA_summary_fcn}
\title{Summarise Osteoarthritis (OA) Data}
\usage{
OA_summary_fcn(am_all)
}
\arguments{
\item{am_all}{A data.frame representing the full simulation output.}
}
\value{
A data.frame with the summarised OA prevalence statistics.
}
\description{
Processes the full simulation output to calculate the prevalence of OA,
stratified by year, age group, and sex.
}
