% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OA_update_fcn.R
\name{OA_update}
\alias{OA_update}
\title{Update Osteoarthritis (OA) Status}
\usage{
OA_update(am_curr, am_new, cycle.coefficents, OA_cust)
}
\arguments{
\item{am_curr}{A data.table representing the attribute matrix for the current
cycle. It contains the state of the population before the update.}

\item{am_new}{A data.table representing the attribute matrix for the next
cycle. This is where the updated OA status will be stored.}

\item{cycle.coefficents}{A list or data.frame of model coefficients for the
OA initiation and progression equations, which also contains the SF-6D
utility decrements.}

\item{OA_cust}{A data.frame with customisation factors for OA coefficients.}
}
\value{
A list containing two data.tables:
\item{am_curr}{The input \code{am_curr} with intermediate calculations and an
updated \code{sf6d_change} column.}
\item{am_new}{The \code{am_new} data.table with updated OA status (\code{oa}, \code{kl2},
\code{kl3}, \code{kl4}, \code{kl_score}).}
}
\description{
This function models the initiation and progression of osteoarthritis (OA) for
one simulation cycle. It calculates the probability of developing OA (KL grade 2)
and progressing to KL grades 3 and 4 based on various risk factors.
}
