% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_optimization.R
\name{optimize_simulation_attributes}
\alias{optimize_simulation_attributes}
\title{Optimize Simulation Attributes}
\usage{
optimize_simulation_attributes(attributes, compress_flags = TRUE)
}
\arguments{
\item{attributes}{Matrix of simulation attributes}

\item{compress_flags}{Whether to compress factor levels}
}
\value{
Memory-optimized attribute matrix
}
\description{
Optimizes attribute matrices for memory efficiency during simulations.
}
