% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_processing.R
\name{parallel_bootstrap}
\alias{parallel_bootstrap}
\title{Parallel Bootstrapping for Uncertainty Analysis}
\usage{
parallel_bootstrap(data, n_boot = 1000, statistic, n_cores = NULL, ...)
}
\arguments{
\item{data}{Original dataset to bootstrap}

\item{n_boot}{Number of bootstrap samples}

\item{statistic}{Function to compute on each bootstrap sample}

\item{n_cores}{Number of CPU cores to use}

\item{...}{Additional arguments to pass to the statistic function}
}
\value{
List of bootstrap results
}
\description{
Performs bootstrap resampling in parallel for uncertainty analysis.
}
