% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_processing.R
\name{parallel_cohort_analysis}
\alias{parallel_cohort_analysis}
\title{Parallel Cohort Analysis}
\usage{
parallel_cohort_analysis(
  cohort_definitions,
  base_data,
  analysis_function,
  n_cores = NULL
)
}
\arguments{
\item{cohort_definitions}{List of cohort definitions (filters and parameters)}

\item{base_data}{Base population data to filter}

\item{analysis_function}{Function to run on each cohort}

\item{n_cores}{Number of cores to use}
}
\value{
Results for each cohort
}
\description{
Analyzes multiple population cohorts in parallel.
}
