% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_processing.R
\name{parallel_cost_calculation}
\alias{parallel_cost_calculation}
\title{Parallel Cost Calculation}
\usage{
parallel_cost_calculation(population_data, cost_configs, n_cores = NULL)
}
\arguments{
\item{population_data}{Large population dataset}

\item{cost_configs}{List of cost configurations to apply}

\item{n_cores}{Number of cores to use}
}
\value{
Combined cost calculation results
}
\description{
Calculates costs in parallel to speed up processing for large populations.
}
