% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_processing.R
\name{parallel_internal_function}
\alias{parallel_internal_function}
\title{Parallel Processing Wrapper for Internal Functions}
\usage{
parallel_internal_function(fun_name, data_list, n_cores = NULL, ...)
}
\arguments{
\item{fun_name}{Name of the function to run in parallel}

\item{data_list}{List of data to process in parallel}

\item{n_cores}{Number of cores to use}

\item{...}{Additional args to pass to the function}
}
\value{
Results from parallel processing
}
\description{
Provides parallel versions of internally-used functions where appropriate.
}
