% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_processing.R
\name{parallel_psa}
\alias{parallel_psa}
\title{Parallel PSA (Probabilistic Sensitivity Analysis)}
\usage{
parallel_psa(param_samples, base_config, n_cores = NULL, run_function = NULL)
}
\arguments{
\item{param_samples}{Matrix/data frame of parameter samples from distributions}

\item{base_config}{Base configuration to modify for each sample}

\item{n_cores}{Number of CPU cores to use}

\item{run_function}{Function to run for each parameter set}
}
\value{
Results for each parameter combination
}
\description{
Runs probabilistic sensitivity analysis in parallel across parameter sets.
}
