% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_analytics.R
\name{perform_advanced_feature_selection}
\alias{perform_advanced_feature_selection}
\title{Advanced Feature Selection}
\usage{
perform_advanced_feature_selection(data, target, config, method = "boruta")
}
\arguments{
\item{data}{Feature matrix}

\item{target}{Target variable}

\item{config}{Configuration}

\item{method}{Feature selection method}
}
\value{
Selected features
}
\description{
Advanced Feature Selection
}
