% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_analytics.R
\name{perform_dimensionality_reduction}
\alias{perform_dimensionality_reduction}
\title{Dimensionality Reduction}
\usage{
perform_dimensionality_reduction(
  data,
  config,
  method = "pca",
  n_components = 2
)
}
\arguments{
\item{data}{Data matrix}

\item{config}{Configuration}

\item{method}{Reduction method}

\item{n_components}{Number of components}
}
\value{
Reduced dimensionality data
}
\description{
Dimensionality Reduction
}
