% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_analytics.R
\name{perform_patient_clustering}
\alias{perform_patient_clustering}
\title{Patient Clustering and Stratification}
\usage{
perform_patient_clustering(patient_data, config, method = "kmeans")
}
\arguments{
\item{patient_data}{Patient data matrix}

\item{config}{Advanced analytics configuration}

\item{method}{Clustering method to use}
}
\value{
Clustering results
}
\description{
Patient Clustering and Stratification
}
