% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcement_learning.R
\name{personalized_policy_learning}
\alias{personalized_policy_learning}
\title{Personalized Treatment Policy Learning}
\usage{
personalized_policy_learning(patient_data, environment, config)
}
\arguments{
\item{patient_data}{Patient data with features}

\item{environment}{Treatment environment}

\item{config}{RL configuration}
}
\value{
Personalized policies
}
\description{
Personalized Treatment Policy Learning
}
