% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pji_module.R
\name{pji_module}
\alias{pji_module}
\title{Main PJI Module Function}
\usage{
pji_module(am_curr, am_new, pji_params)
}
\arguments{
\item{am_curr}{Current attribute matrix}

\item{am_new}{Next cycle attribute matrix}

\item{pji_params}{Complete PJI parameter set}
}
\value{
Updated attribute matrices with PJI modeling
}
\description{
Main PJI Module Function
}
