% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_visualization.R
\name{plot_convergence_diagnostics}
\alias{plot_convergence_diagnostics}
\title{Plot Convergence Diagnostics}
\usage{
plot_convergence_diagnostics(
  psa_results,
  title = "PSA Convergence Diagnostics"
)
}
\arguments{
\item{psa_results}{PSA results with convergence assessment}

\item{title}{Plot title}
}
\value{
ggplot object or list of plots
}
\description{
Plot Convergence Diagnostics
}
