% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_visualization.R
\name{plot_nmb_distribution}
\alias{plot_nmb_distribution}
\title{Plot Net Monetary Benefit (NMB) Distribution}
\usage{
plot_nmb_distribution(nmb_stats, title = "Net Monetary Benefit Distribution")
}
\arguments{
\item{nmb_stats}{NMB statistics from calculate_nmb()}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Plot Net Monetary Benefit (NMB) Distribution
}
