% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_visualization.R
\name{plot_parameter_correlations}
\alias{plot_parameter_correlations}
\title{Plot Parameter Correlation Matrix}
\usage{
plot_parameter_correlations(
  correlation_data,
  title = "Parameter Correlation Matrix",
  subtitle = "Pearson Correlations with ICER"
)
}
\arguments{
\item{correlation_data}{Correlation data from analyze_parameter_correlations()}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}
}
\value{
ggplot object
}
\description{
Plot Parameter Correlation Matrix
}
