% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_visualization.R
\name{plot_tornado_diagram}
\alias{plot_tornado_diagram}
\title{Plot Tornado Diagram for Parameter Influence}
\usage{
plot_tornado_diagram(
  influence_data,
  top_n = 10,
  title = "Parameter Influence (Tornado Diagram)",
  subtitle = "Impact of Parameter Uncertainty on ICER"
)
}
\arguments{
\item{influence_data}{Parameter influence data from analyze_parameter_influence()}

\item{top_n}{Number of top parameters to display}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}
}
\value{
ggplot object
}
\description{
Plot Tornado Diagram for Parameter Influence
}
