% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_visualization.R
\name{plot_voi_analysis}
\alias{plot_voi_analysis}
\title{Plot Value of Information (VOI) Analysis}
\usage{
plot_voi_analysis(voi_stats, title = "Value of Information Analysis")
}
\arguments{
\item{voi_stats}{VOI statistics from calculate_voi()}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Plot Value of Information (VOI) Analysis
}
