% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_monitoring.R
\name{profile_memory}
\alias{profile_memory}
\title{Benchmark function with memory profiling}
\usage{
profile_memory(func, ...)
}
\arguments{
\item{func}{The function to benchmark}

\item{...}{Arguments to pass to the function}
}
\value{
Profiled memory usage
}
\description{
Benchmark function with memory profiling
}
