% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_optimization.R
\name{profile_memory_usage}
\alias{profile_memory_usage}
\title{Memory Usage Profiler}
\usage{
profile_memory_usage(expr, interval = 0.01)
}
\arguments{
\item{expr}{Expression to profile}

\item{interval}{Sampling interval for memory profiler}
}
\value{
Memory profiling results
}
\description{
Profiles memory usage of AUS-OA functions.
}
