% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_private_pathways_module.R
\name{public_private_pathways_module}
\alias{public_private_pathways_module}
\title{Main Public-Private Pathways Module Function}
\usage{
public_private_pathways_module(am_curr, am_new, pathway_params)
}
\arguments{
\item{am_curr}{Current attribute matrix}

\item{am_new}{Next cycle attribute matrix}

\item{pathway_params}{Parameters from configuration}
}
\value{
List containing updated matrices and pathway summary
}
\description{
Main Public-Private Pathways Module Function
}
