% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_calibration.R
\name{quantify_parameter_uncertainty}
\alias{quantify_parameter_uncertainty}
\title{Uncertainty Quantification for Parameters}
\usage{
quantify_parameter_uncertainty(bayesian_model, parameters = NULL)
}
\arguments{
\item{bayesian_model}{Fitted Bayesian model}

\item{parameters}{Parameters of interest}
}
\value{
Uncertainty quantification results
}
\description{
Uncertainty Quantification for Parameters
}
