% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_analytics.R
\name{recognize_treatment_patterns}
\alias{recognize_treatment_patterns}
\title{Pattern Recognition in Treatment Trajectories}
\usage{
recognize_treatment_patterns(longitudinal_data, config)
}
\arguments{
\item{longitudinal_data}{Longitudinal patient data}

\item{config}{Configuration}
}
\value{
Pattern recognition results
}
\description{
Pattern Recognition in Treatment Trajectories
}
