% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions_fcn.R
\name{revisions_fcn}
\alias{revisions_fcn}
\title{Calculate TKA Revision Risk}
\usage{
revisions_fcn(cycle.coefficents, am_curr)
}
\arguments{
\item{cycle.coefficents}{A data.frame or list containing the model
coefficients for the revision model, which must have names starting with "cr_".}

\item{am_curr}{A data.frame representing the attribute matrix for the current
cycle. It must contain variables for age, sex, BMI, ASA score, public/private
status, and time since primary TKA (\code{agetka1}, \code{agetka2}).}
}
\value{
The updated \code{am_curr} data.frame with new columns for the calculated
revision hazard (\code{rev_haz1}, \code{rev_haz2}) and a binary indicator for whether
a revision occurred in this cycle (\code{revi}).
}
\description{
This function calculates the hazard of a total knee arthroplasty (TKA)
revision for each individual in the attribute matrix. The model is based on
a linear predictor for individual risk factors and a spline function to model
the baseline hazard over time since the primary TKA.
}
