% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_processing.R
\name{run_parallel_simulations}
\alias{run_parallel_simulations}
\title{Parallel Simulation Execution}
\usage{
run_parallel_simulations(scenarios, n_cores = NULL, export_vars = NULL)
}
\arguments{
\item{scenarios}{List of scenario configurations to run in parallel}

\item{n_cores}{Number of CPU cores to use}

\item{export_vars}{Additional variables to export to parallel workers}
}
\value{
List of simulation results for each scenario
}
\description{
Runs multiple simulation scenarios in parallel to speed up analysis.
}
