% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_monitoring.R
\name{run_performance_benchmarks}
\alias{run_performance_benchmarks}
\title{Run comprehensive performance benchmark}
\usage{
run_performance_benchmarks(iterations = 3)
}
\arguments{
\item{iterations}{Number of iterations for each test}
}
\value{
List of all benchmark results
}
\description{
Executes performance tests on all main functions
}
