% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_framework.R
\name{run_psa_simulation}
\alias{run_psa_simulation}
\title{Run PSA Simulation}
\usage{
run_psa_simulation(parameter_matrix, simulation_function, ...)
}
\arguments{
\item{parameter_matrix}{Matrix of parameter sets from PSA sampling}

\item{simulation_function}{Function to run simulation with given parameters}

\item{...}{Additional arguments to pass to simulation function}
}
\value{
List containing PSA results and diagnostics
}
\description{
Run PSA Simulation
}
