% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_framework.R
\name{sample_parameters}
\alias{sample_parameters}
\title{Sample Parameters from Distributions}
\usage{
sample_parameters(distributions, n_samples, seed = NULL)
}
\arguments{
\item{distributions}{List of parameter distributions}

\item{n_samples}{Number of parameter sets to sample}

\item{seed}{Random seed for reproducibility}
}
\value{
Matrix of sampled parameter values (n_samples x n_parameters)
}
\description{
Sample Parameters from Distributions
}
