% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcement_learning.R
\name{sarsa_treatment}
\alias{sarsa_treatment}
\title{SARSA Algorithm for Treatment Optimization}
\usage{
sarsa_treatment(environment, config)
}
\arguments{
\item{environment}{Treatment environment}

\item{config}{RL configuration}
}
\value{
Learned policy
}
\description{
SARSA Algorithm for Treatment Optimization
}
