% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pji_module.R
\name{simulate_pji_events}
\alias{simulate_pji_events}
\title{Simulate PJI Events}
\usage{
simulate_pji_events(am_curr)
}
\arguments{
\item{am_curr}{Current attribute matrix with PJI risk probabilities}
}
\value{
Updated attribute matrix with PJI events
}
\description{
Simulate PJI Events
}
