% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TKA_update_fcn.R
\name{TKA_update_fcn}
\alias{TKA_update_fcn}
\title{Update Total Knee Arthroplasty (TKA) Status}
\usage{
TKA_update_fcn(
  am_curr,
  am_new,
  cycle.coefficents,
  TKR_cust,
  summary_TKR_observed_diff
)
}
\arguments{
\item{am_curr}{A data.table representing the attribute matrix for the current
cycle.}

\item{am_new}{A data.table representing the attribute matrix for the next
cycle.}

\item{cycle.coefficents}{A list or data.frame of model coefficients.}

\item{TKR_cust}{A data.frame with customisation factors for TKA coefficients.}

\item{summary_TKR_observed_diff}{A data.frame containing the difference
between observed and expected TKA rates.}
}
\value{
A list containing the updated attribute matrices and a summary of TKA
risk.
}
\description{
This function models the probability of receiving a TKA for one simulation
cycle.
}
