% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pros_fcn.R
\name{update_pros_fcn}
\alias{update_pros_fcn}
\title{Update Patient-Reported Outcomes (PROs)}
\usage{
update_pros_fcn(am_new, cycle.coefficents)
}
\arguments{
\item{am_new}{A data.table representing the attribute matrix with the
latest events for the current cycle.}

\item{cycle.coefficents}{A list of model coefficients.}
}
\value{
The `am_new` data.table with updated `pain` and `function_score` columns.
}
\description{
This function updates the pain and function scores for each individual in
the attribute matrix based on events in the current cycle.
}
