% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_integration.R
\name{validate_psa_results}
\alias{validate_psa_results}
\title{Validate PSA Results}
\usage{
validate_psa_results(psa_results)
}
\arguments{
\item{psa_results}{PSA results to validate}
}
\value{
Validation summary
}
\description{
Validate PSA Results
}
