% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waiting_list_module.R
\name{waiting_list_module}
\alias{waiting_list_module}
\title{Main Waiting List Module Function}
\usage{
waiting_list_module(am_curr, am_new, waiting_list_params)
}
\arguments{
\item{am_curr}{Current attribute matrix}

\item{am_new}{Next cycle attribute matrix}

\item{waiting_list_params}{Parameters from configuration}
}
\value{
List containing updated matrices and waiting list summary
}
\description{
Main Waiting List Module Function
}
