% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_integration.R
\name{write_psa_summary_text}
\alias{write_psa_summary_text}
\title{Write PSA Summary to Text File}
\usage{
write_psa_summary_text(report, filename)
}
\arguments{
\item{report}{PSA report}

\item{filename}{Output filename}
}
\description{
Write PSA Summary to Text File
}
