% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_costs_fcn.R
\name{calculate_costs_fcn}
\alias{calculate_costs_fcn}
\title{Calculate Costs for a Simulation Cycle}
\usage{
calculate_costs_fcn(am_new, costs_config)
}
\arguments{
\item{am_new}{A data.table representing the attribute matrix with the
latest events for the current cycle.}

\item{costs_config}{A list containing the detailed cost parameters from the
model configuration file.}
}
\value{
The \code{am_new} data.table with new columns for costs incurred
during the cycle, broken down by perspective.
}
\description{
This function calculates various costs for each individual in the attribute
matrix based on events that occurred in the current cycle and ongoing states.
It uses a detailed, component-based cost configuration.
}
